Moved waitForVisible to wait folder
